move search to new file
